+++ src/ints/int10_modes.cpp
@@ -791,8 +791,8 @@ static void FinishSetMode(bool clearmem) {
 #if defined(USE_TTF)
     if (TTF_using() && CurMode->type==M_TEXT) {
         if (ttf.inUse) {
-            ttf.cols = CurMode->twidth;
-            ttf.lins = CurMode->theight;
+            ttf.cols = (int)CurMode->twidth;
+            ttf.lins = (int)CurMode->theight;
         } else if (ttf.lins && ttf.cols && !IS_PC98_ARCH && !IS_VGA_ARCH && CurMode->mode == 3) {
             CurMode->twidth = ttf.cols;
             CurMode->theight = ttf.lins;
